pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "investormate"
version = "0.1.0"
description = "AI-powered stock analysis package combining data, technical indicators, and multi-provider AI analysis"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["finance", "stocks", "ai", "analysis", "yfinance", "trading", "investment", "technical-analysis"]
authors = [
    {name = "InvestorMate Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "yfinance>=0.2.40",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pypdf>=5.0.0",
    "validators>=0.20.0",
    "pandas-ta>=0.3.14b",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-genai>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "build>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/investormate/investormate"
Documentation = "https://github.com/investormate/investormate#readme"
Repository = "https://github.com/investormate/investormate"
"Bug Tracker" = "https://github.com/investormate/investormate/issues"

[tool.setuptools]
packages = ["investormate", "investormate.core", "investormate.ai", "investormate.data", "investormate.analysis", "investormate.documents", "investormate.utils"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=investormate --cov-report=term-missing --cov-report=html"
